Release and Sprint Creation Complete
Release page completed for creating new releases and sprints.
The release calculates the number of sprints to be created including
start and end dates by looking at the sprint length and release date.

Sprint page still left to do

+++ taskRemove.php
@@ -6,12 +6,14 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="stylesheet" href="css/bootstrap.min.css">
   <link rel="stylesheet" href="css/style.css">
-  <script src="js/jquery-1.12.0.js"></script>
+  <script src="js/jquery-2.2.0.js"></script>
   <script src="js/validator.js"></script>
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.0/jquery.min.js"></script>
   <script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
 </head>
 <body>
+
+<!-- Navigation Bar -->
 
 <nav class="navbar navbar-inverse navbar-fixed-top">
   <div class="container-fluid">
@@ -37,8 +39,8 @@
         <li class="dropdown">
           <a class="dropdown-toggle" data-toggle="dropdown" href="#">Planning<span class="caret"></span></a>
           <ul class="dropdown-menu">
-            <li><a href="releasePlanning.html">Feature/Release Planning</a></li>
-            <li><a href="sprintPlanning.html">Sprint Planning</a></li>
+            <li><a href="releasePlanning.php">Feature/Release Planning</a></li>
+            <li><a href="sprintPlanning.php">Sprint Planning</a></li>
           </ul>
         </li>
         <li><a href="taskboard.html">Task Board</a></li>
@@ -50,31 +52,39 @@
     </div>
   </div>
 </nav>
-  
+
+<!-- Main Container -->
+
 <div class="container">
+
+<!-- PHP Code - 1.Remove Task by the given id. -->
+
   <?php
-    $conn = new mysqli('localhost', 'root', '', 'tempdb');
-    if($conn->connect_errno > 0)
+    if(isset($_GET["id"]))
     {
-      die('Unable to connect to database [' . $conn->connect_error . ']');
-    }
-    $sql = 'DELETE FROM tablefive WHERE id = ' . $_GET["id"] .'';
+      $conn = new mysqli('localhost', 'root', '', 'tempdb');
+      if($conn->connect_errno > 0)
+      {
+        die('Unable to connect to database [' . $conn->connect_error . ']');
+      }
+      $sql = 'DELETE FROM tablefive WHERE id = ' . $_GET["id"] .'';
       if ($conn->query($sql) === TRUE) 
       {
-       echo "<div class='alert alert-success'>
-       <strong>Success!</strong> Task Successfully Removed.
-       </div>";
+       echo "<div class='alert alert-success'><strong>Success!</strong> Task has been successfully removed.</div>";
       } 
       else 
       {
-        echo "<div class='alert alert-failure'>
-        <strong>Error!</strong> " . $sql . "<br>" . $conn->error . "</div>";
+        echo "<div class='alert alert-failure'><strong>Error!</strong> " . $sql . "<br>" . $conn->error . "</div>";
       }
       $conn->close();
-    ?>
+    }
+  ?>
   <h3>Task Backlog</h3>
+
+<!-- List of Tasks in Backlog -->
+
   <div class="table-responsive">        
-    <table class="table">
+    <table class="table table-striped">
       <thead>
         <tr>
           <th>Id</th>
@@ -87,6 +97,9 @@
         </tr>
       </thead>
       <tbody>
+
+<!-- PHP Code - 1.Grab list of Tasks from the database. -->
+
         <?php
           $conn = new mysqli('localhost', 'root', '', 'tempdb');
           if($conn->connect_errno > 0)
@@ -98,14 +111,14 @@
           {
             ?>
               <tr>
-                <td><?php echo $row['id']?></td>
-                <td><?php echo $row['taskName']?></td>
-                <td><?php echo $row['taskDescription']?></td>
-                <td><?php echo $row['taskPriority']?></td>
-                <td><?php echo $row['taskEstimation']?></td>
-                <td><?php echo $row['story_id']?></td>
+                <td><?php echo $row['id'];?></td>
+                <td><?php echo $row['taskName'];?></td>
+                <td><?php echo $row['taskDescription'];?></td>
+                <td><?php echo $row['taskPriority'];?></td>
+                <td><?php echo $row['taskEstimation'];?></td>
+                <td><?php echo $row['story_id'];?></td>
                 <td>
-                  <a class="btn btn-danger" id="removeButton" href="taskRemove.php?id=<?php echo $row['id']?>">
+                  <a class="btn btn-danger" id="removeButton" href="taskRemove.php?id=<?php echo $row['id'];?>">
                     Remove <span class="glyphicon glyphicon-remove"></span>
                   </a>
                 </td>
@@ -119,6 +132,9 @@
 </div>
 
 <div class="container">
+
+<!-- Form for creating new Task -->
+
   <h3> Create New Task </h3>
   <form class="form-horizontal col-lg-8 col-lg-offset-2" id="taskCreationForm" data-toggle="validator" role="form" novalidate="true" action="taskCreate.php" method="post">
     <div class="row form-group has-feedback">
@@ -128,7 +144,7 @@
     </div>
     <div class="row form-group has-feedback">
       <label class="control-label" for="task_description">Task Description:</label>
-      <textarea type="text" class="form-control" name="task_description" maxlength="100" placeholder="Enter Task Description" rows="3" required></textarea>
+      <textarea type="text" class="form-control" name="task_description" maxlength="1000" placeholder="Enter Task Description" rows="3" required></textarea>
       <span class="glyphicon form-control-feedback" aria-hidden="true"></span>
     </div>
     <div class="row">
@@ -145,7 +161,7 @@
           while($row = mysqli_fetch_array($sql))          
           {
             ?> 
-            <option><?php echo $row['id'] . '. ' . $row['storyName']?></option>
+            <option><?php echo $row['id'] . '. ' . $row['storyName'];?></option>
             <?php
           }
           ?>
